refactor(star-adventurer-gti): extract enable_sidereal_tracking_ra helper
Closes #262. Collapse three duplicated `:G1 TRACKING` + `:I1 sidereal_period`
+ `:J1` sequences into a single `enable_sidereal_tracking_ra(transport,
params)` helper in `mount_device/slew.rs`, alongside the existing wire
helpers (`stop_axis_and_wait`, `issue_slew_axis`, `pickup_reslew_axis`).

The helper returns `Result<(), StarAdvError>` and lets each call site
pick its own policy:

- `Telescope::set_tracking(true)` (telescope.rs): maps to `ASCOMError`
  and propagates via `Self::ascom`.
- `spawn_slew_completion_watcher` post-slew restore (watchers.rs): logs
  at `warn` on failure; flips `state.tracking_requested = true` only on
  success — preserves the previous "only if the final StartMotion
  succeeds" rule.
- `spawn_pulse_guide_watcher` RA-only restore (watchers.rs): logs at
  `warn` on failure.

Wire frames on the happy path are unchanged: same three commands in the
same order. On failure, both watchers now short-circuit on the first
failing send. The pulse-guide watcher already short-circuited; the slew
watcher previously tried all three regardless and emitted up to three
warnings — almost certainly accidental, since a SetMotionMode failure
guarantees the subsequent SetStepPeriod / StartMotion fail with the
same underlying error.

Spotted during the issue-253 refactor (PR #261); deliberately deferred
so the structural move stayed clean.

All 308 unit/integration tests still pass; `cargo rail run --profile
commit -q` and `cargo fmt --check` are clean.
